Reorganised folder structure
created seperate folders for each element of the archiver.
Retriever, explorer and Downloader.

Moved util functionality to util folder and db functionality to db folder.
