+++ pipeline/archiver/retriever/appIconRetriever.py
