--- pipeline/scraper/json_dumps/whatsapp_links/whatsapp_all_link_href.json
@@ -1,19 +1,0 @@
-[
-  "//lh6.ggpht.com/mp86vbELnqLi2FzvhiKdPX31_oiTRLNyeK8x4IIrbF5eD1D5RdnVwjQP0hwMNR_JdA=w300",
-  "//lh5.ggpht.com/0VYAvZLR9YhosF-thqm8xl8EWsCfrEY_uk2og2f59K8IOx5TfPsXjFVwxaHVnUbuEjc=w170",
-  "//lh3.googleusercontent.com/ZZPdzvlpK9r_Df9C3M7j1rNRi7hhHRvPhlklJ3lfi5jk86Jd1s0Y5wcQ1QgbVaAP5Q=w170",
-  "//lh3.googleusercontent.com/tm_N1osJGfifuRlMfEip4kZFD5QCtd42CYYEhUsxABzIoEn6Nb9UXjRVmjoeKvMFUCzF=w170",
-  "//lh3.googleusercontent.com/8WsuI2NpvYJljZshzCgLXKRLnFLzz3q4I3JGw3L0vfintPGEiTChOXDIrnoXFFfme58=w170",
-  "//lh3.googleusercontent.com/nYhPnY2I-e9rpqnid9u9aAODz4C04OycEGxqHG5vxFnA35OGmLMrrUmhM9eaHKJ7liB-=w170",
-  "//lh3.googleusercontent.com/aYbdIM1abwyVSUZLDKoE0CDZGRhlkpsaPOg9tNnBktUQYsXflwknnOn2Ge1Yr7rImGk=w170",
-  "//lh6.ggpht.com/8-N_qLXgV-eNDQINqTR-Pzu5Y8DuH0Xjz53zoWq_IcBNpcxDL_gK4uS_MvXH00yN6nd4=w170",
-  "//lh3.googleusercontent.com/dH4iXrVFkT-6s62wew4Hg4E5K7S8eeJHs53L9UjPHFJmZULw51aaBUkUrm1LvQ79Lnk=w170",
-  "//lh3.googleusercontent.com/l-ZZOFGyeKYz3stUbxTECHYnXcRD66C9g0tjiWA_okVIxZyb0E7_esU8LRpq_0LFCu8Y=w170",
-  "//lh6.ggpht.com/653qJf6Ro417FXETOBl3cAJ9hborY6PnCONzLBapvNDw6VgRximFc9YKWDGc82yKKNEV=w170",
-  "//lh3.googleusercontent.com/Pwib7QDQnb-zK756VWLL5SoMyorKm6nsjImGay3bH4QwYVRjU1G8inXuQWR4r0ZfIxY=w170",
-  "//lh3.googleusercontent.com/jVXglyWWL5J2y1vRN-7Jy3_ozvvZc4w5486IAkbAIrWcNN_vn7YuIvhc1JDtGq43BqGl=w170",
-  "//lh3.googleusercontent.com/sRmOaGKtn9ByTyHyEjPHLOqcgSdROfqiF7-q3P0B_8AkPywuZm-CDt1pX9z8Oob4Y3Od=w170",
-  "//lh3.googleusercontent.com/QfAEt_ya6-n8w_TD9-PsghFC2DMSO7fLGNZB4cQ3RtbBbHFkXJE_gxOc3l32-j6LXg=w170",
-  "//lh3.googleusercontent.com/CrsX4NgCfgY1B8ry-cP0qdEgR7cB7717qmVajenGgWuh5t7jCg6cRQxLi1c73hj3iQ=w170",
-  "https://lh4.ggpht.com/7jfcnc_BKw-oGhC-aFOCQoO6cgAHXi5daP85zSzvOBGGv7FHCHL4f7oJjgswVCoXqsRp=w170"
-]
