--- pipeline/scraper/json_dumps/whatsapp_links/whatsapp_all_link_json.json
@@ -1,1 +1,0 @@
-{}
