--- pipeline/scraper/json_dumps/whatsapp_links/whatsapp_local_link_json.json
@@ -1,1 +1,0 @@
-{}
